+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+- Add option to add include polyfills in build.
 
 ## [1.3.1]
 - Bumps ember-cli-qunit from 4.3.0 to 4.3.2.